+++ Causal_Web/Causal_Web.pyproj
@@ -27,30 +27,30 @@
     <Compile Include="database\run_meta.py" />
     <Compile Include="database\__init__.py" />
     <Compile Include="engine\base.py" />
-    <Compile Include="engine\bridge.py" />
+    <Compile Include="engine\models\bridge.py" />
     <Compile Include="engine\causal_analyst.py" />
     <Compile Include="engine\explanation_rules.py" />
-    <Compile Include="engine\graph.py" />
-    <Compile Include="engine\logger.py" />
-    <Compile Include="engine\logging_models.py" />
+    <Compile Include="engine\models\graph.py" />
+    <Compile Include="engine\logging\logger.py" />
+    <Compile Include="engine\models\logging.py" />
     <Compile Include="engine\meta_node.py" />
-    <Compile Include="engine\node.py" />
-    <Compile Include="engine\node_manager.py" />
-    <Compile Include="engine\observer.py" />
+    <Compile Include="engine\models\node.py" />
+    <Compile Include="engine\tick_engine\node_manager.py" />
+    <Compile Include="engine\models\observer.py" />
     <Compile Include="engine\services\serialization_service.py" />
     <Compile Include="engine\services\node_services.py" />
     <Compile Include="engine\services\sim_services.py" />
     <Compile Include="engine\services\__init__.py" />
-    <Compile Include="engine\tick.py" />
-    <Compile Include="engine\log_interpreter.py" />
+    <Compile Include="engine\models\tick.py" />
+    <Compile Include="engine\logging\log_interpreter.py" />
     <Compile Include="engine\tick_engine\bridge_manager.py" />
     <Compile Include="engine\tick_engine\core.py" />
     <Compile Include="engine\tick_engine\evaluator.py" />
-    <Compile Include="engine\tick_engine\log_utils.py" />
+    <Compile Include="engine\logging\log_utils.py" />
     <Compile Include="engine\tick_engine\orchestrators.py" />
     <Compile Include="engine\tick_engine\__init__.py" />
-    <Compile Include="engine\tick_router.py" />
-    <Compile Include="engine\tick_seeder.py" />
+    <Compile Include="engine\tick_engine\tick_router.py" />
+    <Compile Include="engine\tick_engine\tick_seeder.py" />
     <Compile Include="engine\__init__.py" />
     <Compile Include="graph\io.py" />
     <Compile Include="graph\model.py" />
