+++ Causal_Web/__init__.py
@@ -1,5 +1,5 @@
 """Causal_Web package initialization."""
 
-from .engine.node_manager import NodeManager
+from .engine.tick_engine.node_manager import NodeManager
 
 __all__ = ["NodeManager"]
