+++ Causal_Web/engine/explanation_rules.py
@@ -5,7 +5,7 @@ import os
 from typing import Dict, List
 
 from .causal_analyst import ExplanationEvent, CausalAnalyst
-from .base import OutputDirMixin
+from .models.base import OutputDirMixin
 
 
 class ExplanationRuleMatcher(OutputDirMixin):
