+++ tests/test_clusters.py
@@ -1,6 +1,6 @@
 import math
-from Causal_Web.engine.graph import CausalGraph
-from Causal_Web.engine.node import Node
+from Causal_Web.engine.models.graph import CausalGraph
+from Causal_Web.engine.models.node import Node
 
 
 def test_hierarchical_cluster_assignment():
